+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2026-04-23]
+## [1.0.0 - 2026-04-27]
 ### Added
 - First version of the project
 - Spring Application
@@ -259,6 +259,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Description document and related services
 - Cohort definition in a query
 - Send Query To Bridgehead Event
+- Send Query To Bridgehead Event scheduled job
 
 
 ### Changed